refactor(invoice-inbox): export uploadAndExtract from lib
Pure move of uploadAndExtract and its private helpers (sanitisers,
page-count/slice, sandbox check, MIME/size consts) from index.ts into
lib/upload-and-extract.ts so a future channel extension can import the
shared funnel the way document-extraction already imports
extract-invoice-fields. No behavior change; existing tests prove it.
